Removed A* pathfinder utility, NinjaBot small update
