+++ bots/spiteful_bot.py
@@ -1,10 +1,9 @@
 # Maintainer: ninja
 import random
-from collections import namedtuple
+from collections import namedtuple, defaultdict
 from bots import BaseBot
 from util.hexagon import Hex, DIAGONALS, is_hex
 from util.settings import Settings
-from util.pathfinding import a_star
 from api.actions import Move, Push, Idle
 
 
@@ -43,7 +42,7 @@ class SpitefulBot(BaseBot):
 
     def get_action(self, wi):
         self.update(wi)
-        if self.ap < 10:
+        if self.ap < Move.ap:
             self.logger(f'Out of AP.')
             return Idle()
 
@@ -66,7 +65,7 @@ class SpitefulBot(BaseBot):
             path = self.path_as_close(self.map_center)
             if path:
                 trunc_path = []
-                moves = int(self.ap/10)
+                moves = int(self.ap/Move.ap)
                 for t in path[:moves]:
                     trunc_path.append(Move(t))
                     if self.map_center.get_distance(t) <= target_distance:
@@ -87,8 +86,8 @@ class SpitefulBot(BaseBot):
         self.logger(f'Idling.')
         return Idle()
 
-    def move_tile_cost(self, tile):
-        is_obstacle = tile in (self.wi.pits | self.wi.walls | set(self.wi.positions))
+    def move_tile_cost(self, origin, target):
+        is_obstacle = target in (self.wi.pits | self.wi.walls | set(self.wi.positions))
         obs_cost = float('inf') if is_obstacle else 0
         return 1 + obs_cost
 
@@ -96,7 +95,7 @@ class SpitefulBot(BaseBot):
         return a_star(self.pos, target, cost=self.move_tile_cost)
 
     def get_paths(self, targets, sort=len):
-        targets = (t for t in targets if self.move_tile_cost(t) < float('inf'))
+        targets = (t for t in targets if self.move_tile_cost(None, t) < float('inf'))
         path_results = (self.get_path(t) for t in targets)
         paths = [p for p in path_results if p is not None]
         if sort:
@@ -121,7 +120,7 @@ class SpitefulBot(BaseBot):
         assert pit not in self.blocked_pits
         assert pit in enemy.neighbors
         # Check ap and sequence blockers
-        ap_cost = 30
+        ap_cost = Push.ap
         start_pos = next(pit.straight_line(enemy))
         cannot_reach = self.pos.get_distance(start_pos) > self.ap - ap_cost
         if start_pos in self.sequence_blockers or cannot_reach:
@@ -131,7 +130,7 @@ class SpitefulBot(BaseBot):
             start_path = self.get_path(start_pos)
         if start_path is None:
             return None
-        ap_cost += len(start_path) * 10
+        ap_cost += len(start_path) * Move.ap
         if ap_cost > self.ap:
             return None
         # Collect actions and vfx
@@ -158,7 +157,7 @@ class SpitefulBot(BaseBot):
         assert len(mid_points) == 1
         mid_point = mid_points.pop()
         # Check ap and sequence blockers
-        ap_cost = 70
+        ap_cost = Move.ap + Push.ap*2
         vector = enemy - mid_point
         start_pos = enemy + vector
         cannot_reach = self.pos.get_distance(start_pos) > self.ap - ap_cost
@@ -169,7 +168,7 @@ class SpitefulBot(BaseBot):
             start_path = self.get_path(start_pos)
         if start_path is None:
             return None
-        ap_cost += len(start_path) * 10
+        ap_cost += len(start_path) * Move.ap
         if ap_cost > self.ap:
             return None
         # Collect actions and vfx
@@ -197,7 +196,7 @@ class SpitefulBot(BaseBot):
         assert neighbor.get_distance(pit) == 1
         assert neighbor.get_distance(enemy) == 1
         # Check ap and sequence blockers
-        ap_cost = 80
+        ap_cost = Move.ap*2 + Push.ap*2
         start_pos = next(neighbor.straight_line(enemy))
         mid_pos = next(pit.straight_line(neighbor))
         seq_blocked = {start_pos, neighbor, mid_pos} & self.sequence_blockers
@@ -209,7 +208,7 @@ class SpitefulBot(BaseBot):
             start_path = self.get_path(start_pos)
         if start_path is None:
             return None
-        ap_cost += len(start_path) * 10
+        ap_cost += len(start_path) * Move.ap
         if ap_cost > self.ap:
             return None
         # Collect actions and vfx
@@ -263,6 +262,41 @@ class SpitefulBot(BaseBot):
         lethal_sequences = sorted(
             lethal_sequences, key=lambda x: x.ap_cost + len(x.actions) / 100)
         return lethal_sequences
+
+
+def a_star(origin, target, cost):
+    def _get_full_path(node, came_from):
+        full_path = [node]
+        while node in came_from:
+            node = came_from[node]
+            full_path.insert(0, node)
+        return tuple(full_path[1:])
+
+    if origin is target:
+        return None
+
+    open_set = {origin}
+    came_from = {}
+    partial_score = defaultdict(lambda: float('inf'))
+    partial_score[origin] = 0
+    guess_score = defaultdict(lambda: float('inf'))
+    guess_score[origin] = origin.get_distance(target)
+
+    while open_set:
+        best_nodes = sorted(open_set, key=lambda x: guess_score[x])
+        current = best_nodes[0]
+        if current is target:
+            return _get_full_path(current, came_from)
+        open_set.remove(current)
+        for neighbor in current.neighbors:
+            tentative_partial_score = partial_score[current] + cost(current, neighbor)
+            if tentative_partial_score < partial_score[neighbor]:
+                came_from[neighbor] = current
+                partial_score[neighbor] = tentative_partial_score
+                guess_score[neighbor] = tentative_partial_score + neighbor.get_distance(target)
+                if neighbor not in open_set:
+                    open_set.add(neighbor)
+    return None
 
 
 BOT = SpitefulBot
